fix cooking info-* packages when buildlogging is turned on

+++ build/cook.py
@@ -744,7 +744,7 @@ def _cookPackageObject(repos, cfg, recipeClass, sourceVersion, prep=True,
             logFile.close()
         raise
 
-    if logBuild:
+    if logBuild and recipeObj._autoCreatedFileCount:
         logFile.close()
         os.unlink(logPath)
         if cfg.noClean:
